split validator up into multiple source files
